+++ source/src_pdiag/pdiag_double.h
@@ -6,7 +6,7 @@
 #include "../module_base/matrix.h"
 #include "../module_base/complexmatrix.h"
 #include "diag_scalapack_gvx.h"
-#include "src_parallel/parallel_orbitals.h"
+#include "module_orbital/parallel_orbitals.h"
 #include "src_lcao/local_orbital_wfc.h"
 
 class Pdiag_Double
